third_party/blink/web_tests/fast/events/before-unload-return-bad-value-expected.txt

Tests that an exception is thrown when the value returned in the beforeunload callback cannot be converted to a String

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Exception was thrown
PASS testMessage is "Uncaught Exception in toString()"
PASS successfullyParsed is true

TEST COMPLETE
